# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perfect Balance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface ProductSpecs {
  /** @nullable */
  material?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  for?: string | null;
  /** @nullable */
  discipline?: string | null;
}

export interface Product {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  sizes?: string | null;
  /** @nullable */
  size?: string | null;
  package_qty: number;
  /** @nullable */
  price_pair?: number | null;
  price_box: number;
  /** @nullable */
  badge?: string | null;
  description: string;
  specs?: ProductSpecs | null;
  image: string;
}

export interface CartItem {
  productId: number;
  name: string;
  qty: number;
  unit: string;
  price: number;
}

export interface CreateOrderBody {
  name: string;
  phone: string;
  email: string;
  city?: string;
  address?: string;
  transportCompany?: string;
  items: CartItem[];
  total: number;
}

export interface OrderConfirmation {
  orderId: string;
  message: string;
}

export interface ContactBody {
  name: string;
  phone: string;
  email?: string;
  clientType?: string;
  message: string;
}

export type GetProductsParams = {
  category?: string;
};