# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perfect Balance API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all products
 */
export const GetProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
});

export const GetProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  sizes: zod.string().nullish(),
  size: zod.string().nullish(),
  package_qty: zod.number(),
  price_pair: zod.number().nullish(),
  price_box: zod.number(),
  badge: zod.string().nullish(),
  description: zod.string(),
  specs: zod
    .union([
      zod.object({
        material: zod.string().nullish(),
        weight: zod.string().nullish(),
        for: zod.string().nullish(),
        discipline: zod.string().nullish(),
      }),
      zod.null(),
    ])
    .optional(),
  image: zod.string(),
});
export const GetProductsResponse = zod.array(GetProductsResponseItem);

/**
 * @summary Get a single product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  category: zod.string(),
  sizes: zod.string().nullish(),
  size: zod.string().nullish(),
  package_qty: zod.number(),
  price_pair: zod.number().nullish(),
  price_box: zod.number(),
  badge: zod.string().nullish(),
  description: zod.string(),
  specs: zod
    .union([
      zod.object({
        material: zod.string().nullish(),
        weight: zod.string().nullish(),
        for: zod.string().nullish(),
        discipline: zod.string().nullish(),
      }),
      zod.null(),
    ])
    .optional(),
  image: zod.string(),
});

/**
 * @summary Place an order
 */
export const CreateOrderBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  city: zod.string().optional(),
  address: zod.string().optional(),
  transportCompany: zod.string().optional(),
  items: zod.array(
    zod.object({
      productId: zod.number(),
      name: zod.string(),
      qty: zod.number(),
      unit: zod.string(),
      price: zod.number(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Submit contact form
 */
export const SubmitContactBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  clientType: zod.string().optional(),
  message: zod.string(),
});

export const SubmitContactResponse = zod.object({
  message: zod.string(),
});