# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perfect Balance API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  productId: number;
  name: string;
  qty: number;
  unit: string;
  price: number;
}