# lib/api-zod/src/generated/types/contactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perfect Balance API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactBody {
  name: string;
  phone: string;
  email?: string;
  clientType?: string;
  message: string;
}