# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perfect Balance API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";

export interface CreateOrderBody {
  name: string;
  phone: string;
  email: string;
  city?: string;
  address?: string;
  transportCompany?: string;
  items: CartItem[];
  total: number;
}