# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perfect Balance API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpecs } from "./productSpecs";

export interface Product {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  sizes?: string | null;
  /** @nullable */
  size?: string | null;
  package_qty: number;
  /** @nullable */
  price_pair?: number | null;
  price_box: number;
  /** @nullable */
  badge?: string | null;
  description: string;
  specs?: ProductSpecs | null;
  image: string;
}