# lib/api-zod/src/generated/types/productSpecs.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perfect Balance API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductSpecs {
  /** @nullable */
  material?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  for?: string | null;
  /** @nullable */
  discipline?: string | null;
}